feat(music): add time in/remaining to track progress; remove from footer (#72)
