+++ src/lib/styles/global.pcss
@@ -25,9 +25,6 @@ body {
 }
 
 body {
-  transition-duration: 10s;
-  transition-property: background;
-  background: var(--bgColor);
   --previousBackgroundGradient: linear-gradient(
     45deg,
     rgb(3, 2, 20),
@@ -37,22 +34,48 @@ body {
     rgb(0, 6, 0)
   );
   --previousForegroundGradient: linear-gradient(
-      45deg,
-      rgb(3, 2, 20),
-      rgb(0, 0, 21),
-      rgb(39, 19, 26),
-      rgb(0, 0, 28),
-      rgb(0, 6, 0)
-    )
-    &: after {position: absolute; opacity: 0; z-index: -1; transition-duration:
-    10s; transition-property: opacity; inset: 0; min-height: 100%;
-    pointer-events: none; content: '' ; html.dark & {background: linear-gradient(
-      to bottom,
-      var(--skyBaseColor) 5%,
-      transparent
-    ) ;}} .dboard {opacity: 0; z-index: 9; transition-duration: 1s;
-    transition-property: opacity;} &.ready {.dboard {opacity: 1;} &: after {opacity:
-    1;}};
+    45deg,
+    rgb(3, 2, 20),
+    rgb(0, 0, 21),
+    rgb(39, 19, 26),
+    rgb(0, 0, 28),
+    rgb(0, 6, 0)
+  );
+  transition-duration: 10s;
+  transition-property: background;
+  background: var(--bgColor);
+  &:after {
+    position: absolute;
+    opacity: 0;
+    z-index: -1;
+    transition-duration: 10s;
+    transition-property: opacity;
+    inset: 0;
+    min-height: 100%;
+    pointer-events: none;
+    content: '';
+    html.dark & {
+      background: linear-gradient(
+        to bottom,
+        var(--skyBaseColor) 5%,
+        transparent
+      );
+    }
+  }
+  .dboard {
+    opacity: 0;
+    z-index: 9;
+    transition-duration: 1s;
+    transition-property: opacity;
+  }
+  &.ready {
+    .dboard {
+      opacity: 1;
+    }
+    &:after {
+      opacity: 1;
+    }
+  }
 }
 
 .header {
