+++ src/lib/widgets/NowPlaying/now-playing.css
@@ -396,36 +396,65 @@
     flex-direction: row;
     margin: 0;
     width: 100%;
-    height: 0.5rem;
-    overflow: hidden;
-    &::before {
+
+    &-bar {
+      display: flex;
+      right: 0;
+      bottom: 0;
+      left: 0;
+      flex-direction: row;
+      height: 0.5rem;
+      overflow: hidden;
+      &::before {
+        position: absolute;
+        opacity: 0.67;
+        z-index: 3;
+        transition: width 2s linear;
+        background: var(--previousForegroundGradient);
+        background-size: 100vw 100vh;
+        width: var(--percent);
+        height: 100%;
+        content: '';
+      }
+      &::before,
+      .playhead {
+        top: 0;
+      }
+      .playhead {
+        position: relative;
+        mix-blend-mode: normal;
+        mix-blend-mode: difference;
+        transition: left 2s linear;
+        background: var(--previousForegroundGradient);
+        background-position: bottom left;
+        width: 0.25rem;
+        height: 1rem;
+        &:hover,
+        &:focus,
+        &:active {
+          background-color: black;
+          width: 0.75rem;
+        }
+      }
+    }
+    .track-time-now,
+    .track-time-remaining {
+      display: inline-block;
       position: absolute;
+      bottom: 1rem;
       opacity: 0.67;
       z-index: 3;
-      transition: width 2s linear;
       background: var(--previousForegroundGradient);
-      width: var(--percent);
-      height: 100%;
-      content: '';
+      background-size: 100vw 100vh;
+      background-clip: text;
+      -webkit-background-clip: text;
+      color: transparent;
     }
-    &::before,
-    .playhead {
-      top: 0;
+    .track-time-now {
+      left: 1rem;
     }
-    .playhead {
-      position: relative;
-      mix-blend-mode: normal;
-      mix-blend-mode: difference;
-      transition: left 2s linear;
-      background: var(--previousForegroundGradient);
-      width: 0.25rem;
-      height: 1rem;
-      &:hover,
-      &:focus,
-      &:active {
-        background-color: black;
-        width: 0.75rem;
-      }
+    .track-time-remaining {
+      right: 1rem;
     }
   }
 
@@ -651,7 +680,6 @@
         h3.track-time {
           opacity: 1;
           margin: 0;
-          background-position: 70% 50%;
           padding-bottom: 1rem;
         }
       }
